+++ test/units/modules/system/interfaces_file/fixtures/golden_output/default_dhcp.test_no_changes
