+++ test/units/modules/system/interfaces_file/fixtures/golden_output/default_dhcp_add_aggi_up.json
