+++ test/units/modules/system/interfaces_file/fixtures/golden_output/default_dhcp_set_aggi_slaves
