+++ test/units/modules/system/interfaces_file/fixtures/golden_output/servers.com.test_no_changes.json
