+++ test/units/modules/system/interfaces_file/fixtures/golden_output/servers.com_add_and_delete_aggi_up
