+++ test/units/modules/system/interfaces_file/fixtures/golden_output/servers.com_set_aggi_slaves
