+++ test/units/modules/system/interfaces_file/fixtures/input/default_dhcp
